+++ lib/faker/fiction/star_wars/en.ex
@@ -1,4 +1,4 @@
-defmodule Faker.StarWars.En do
+defmodule Faker.Fiction.StarWars.En do
   import Faker, only: [sampler: 2]
 
   @moduledoc """
@@ -10,13 +10,13 @@ defmodule Faker.StarWars.En do
 
   ## Examples
 
-      iex> Faker.StarWars.En.character()
+      iex> Faker.Fiction.StarWars.En.character()
       "Greedo"
-      iex> Faker.StarWars.En.character()
+      iex> Faker.Fiction.StarWars.En.character()
       "Jek Tono Porkins"
-      iex> Faker.StarWars.En.character()
+      iex> Faker.Fiction.StarWars.En.character()
       "Poe Dameron"
-      iex> Faker.StarWars.En.character()
+      iex> Faker.Fiction.StarWars.En.character()
       "R4-P17"
   """
   @spec character() :: String.t()
@@ -116,13 +116,13 @@ defmodule Faker.StarWars.En do
 
   ## Examples
 
-      iex> Faker.StarWars.En.planet()
+      iex> Faker.Fiction.StarWars.En.planet()
       "Mon Cala"
-      iex> Faker.StarWars.En.planet()
+      iex> Faker.Fiction.StarWars.En.planet()
       "Ryloth"
-      iex> Faker.StarWars.En.planet()
+      iex> Faker.Fiction.StarWars.En.planet()
       "Endor"
-      iex> Faker.StarWars.En.planet()
+      iex> Faker.Fiction.StarWars.En.planet()
       "Shili"
   """
   @spec planet() :: String.t()
@@ -194,13 +194,13 @@ defmodule Faker.StarWars.En do
 
   ## Examples
 
-      iex> Faker.StarWars.En.quote()
+      iex> Faker.Fiction.StarWars.En.quote()
       "Congratulations. You are being rescued. Please do not resist."
-      iex> Faker.StarWars.En.quote()
+      iex> Faker.Fiction.StarWars.En.quote()
       "What chance do we have? The question is 'what choice'. Run, hide, plead for mercy, scatter your forces. You give way to an enemy this evil with this much power and you condemn the galaxy to an eternity of submission. The time to fight is now!"
-      iex> Faker.StarWars.En.quote()
+      iex> Faker.Fiction.StarWars.En.quote()
       "Will someone get this big walking carpet out of my way?"
-      iex> Faker.StarWars.En.quote()
+      iex> Faker.Fiction.StarWars.En.quote()
       "To be Jedi is to face the truth, and choose. Give off light, or darkness, Padawan. Be a candle, or the night."
   """
   @spec quote() :: String.t()