+++ test/faker/fiction/pokemon_test.exs
@@ -1,0 +1,6 @@
+defmodule Faker.Fiction.PokemonTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Fiction.Pokemon
+  doctest Faker.Fiction.Pokemon.En
+end