+++ test/faker/fiction/star_wars_test.exs
@@ -1,0 +1,6 @@
+defmodule Faker.Fiction.StarWarsTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Fiction.StarWars
+  doctest Faker.Fiction.StarWars.En
+end