+++ test/faker/fiction/superhero_test.exs
@@ -1,0 +1,6 @@
+defmodule Faker.Fiction.SuperheroTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Fiction.Superhero
+  doctest Faker.Fiction.Superhero.En
+end