+++ test/faker/pokemon_test.exs
@@ -1,6 +1,0 @@
-defmodule Faker.PokemonTest do
-  use ExUnit.Case, async: true
-
-  doctest Faker.Pokemon
-  doctest Faker.Pokemon.En
-end