+++ test/faker/star_wars_test.exs
@@ -1,6 +1,0 @@
-defmodule Faker.StarWarsTest do
-  use ExUnit.Case, async: true
-
-  doctest Faker.StarWars
-  doctest Faker.StarWars.En
-end