+++ test/faker/superhero_test.exs
@@ -1,6 +1,0 @@
-defmodule Faker.SuperheroTest do
-  use ExUnit.Case, async: true
-
-  doctest Faker.Superhero
-  doctest Faker.Superhero.En
-end